+++ app/src/main/java/org/cxct/sportlottery/network/index/login/LoginData.kt
@@ -20,11 +20,11 @@ data class LoginData(
     @Json(name = "nickName")
     val nickName: String?, //昵称
     @Json(name = "platformId")
-    val platformId: Int?, //平台id
+    val platformId: Long?, //平台id
     @Json(name = "rechLevel")
     val rechLevel: String?, //充值层级
     @Json(name = "testFlag")
-    val testFlag: Int?, //是否测试用户（0-正常用户，1-游客，2-内部测试）
+    val testFlag: Long?, //是否测试用户（0-正常用户，1-游客，2-内部测试）
     @Json(name = "token")
     val token: String?,
     @Json(name = "uid")
